chore: Move sheets to new home
- Also rename `View/index.tsx` to `View.tsx`.
